refactor(auth): move token lifecycle methods into TokenSessionService
Move RefreshTokenAsync, RevokeUserTokens, and cookie deletion from
AuthenticationService into TokenSessionService. This removes
ITokenProvider, ICookieService, and HybridCache from
AuthenticationService, leaving it focused on auth orchestration.
